requirements.txt
fastapi
uvicorn[standard]
azure-identity
azure-storage-blob
python-dotenv
python-multipart
openai
